cmd/testbcrypt: move tests to actual test file
